requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
python-dotenv==1.0.1
datasets==3.0.1
sentence-transformers==3.2.1
pinecone-client==4.1.1
requests==2.32.3
numpy==1.26.4
torch==2.2.2 
torchvision==0.17.2 
torchaudio==2.2.2
